+++ packages/client/src/hooks/pagination/usePagination.ts
@@ -26,6 +26,7 @@ import {
   promiseFinally,
   promiseResolve,
   pushItem,
+  setTimeoutFn,
   splice,
   trueValue,
   undefinedValue,
@@ -250,22 +251,30 @@ export default <AG extends AlovaGenerics, ListData extends unknown[]>(
       promiseCatch(fetch(fetchMethod as Method, ...customArgs, undefinedValue), noop);
     }
   };
+
+  const computeIsLastPage = () => {
+    const dataRaw = nestedData.v;
+    if (!dataRaw) {
+      return trueValue;
+    }
+    const statesDataVal = listDataGetter(dataRaw as any[]);
+    const pageVal = page.v;
+    const pageCountVal = pageCount.v;
+    const dataLen = isArray(statesDataVal) ? len(statesDataVal) : 0;
+    return pageCountVal ? pageVal >= pageCountVal : dataLen < pageSize.v;
+  };
+  /**
+   * fix #685
+   * @see {https://github.com/alovajs/alova/issues/685}
+   */
+  const isLastPage = create(computeIsLastPage(), 'isLastPage');
   // If the returned data is smaller than the page size, it is considered the last page.
-  const isLastPage = computed(
-    () => {
-      const dataRaw = nestedData.v;
-      if (!dataRaw) {
-        return trueValue;
-      }
-      const statesDataVal = listDataGetter(dataRaw as any[]);
-      const pageVal = page.v;
-      const pageCountVal = pageCount.v;
-      const dataLen = isArray(statesDataVal) ? len(statesDataVal) : 0;
-      return pageCountVal ? pageVal >= pageCountVal : dataLen < pageSize.v;
-    },
-    [page, pageCount, nestedData, pageSize],
-    'isLastPage'
-  );
+  watch([page, pageCount, nestedData, pageSize], async () => {
+    // the reason why delay is used is that needed to wait for the `loading` state to be updated.
+    setTimeoutFn(() => {
+      isLastPage.v = computeIsLastPage();
+    });
+  });
 
   // Update current page cache
   const updateCurrentPageCache = async () => {
@@ -552,7 +561,7 @@ export default <AG extends AlovaGenerics, ListData extends unknown[]>(
         });
       }
 
-      const isLastPageVal = isLastPage.v;
+      const isLastPageVal = computeIsLastPage();
       const fillingItemsLen = len(fillingItems);
       let isLastEmptyPageInNonAppendMode = false;
       if (fillingItemsLen > 0 || isLastPageVal) {